(GH-30) Add Forge hostname and auth token params
Prior to this commit, there was no ability to specify a custom
Forge URI and/or authentication token. This would impact users who
either had their own Forge instance, were premium content users or
both.

This commit adds the `--forge-hostname` and `--forge-token` opts
to the CLI and passes these to `PuppetForge` on initialisation, if
defined.