(CONT-693) - Implement reusable workflow
Prior to this commit, the version of the gem would need to be manually updated,
pushed and merged into main each time before running release_prep.yml.

This commit introduces a "version" input parameter, that can be specified when
running the workflow which eliminates the need for the above.

The commit also introduces the cat team reusable workflows.